+++ website/src/components/Credits.tsx
@@ -5,11 +5,13 @@ import {
   AnthropicLogo,
   BoltLogo,
   BuilderLogo,
+  CloudflareLogo,
   CodeRabbitLogo,
   DigitalOceanLogo,
   HdmLogo,
   OpenAILogo,
   PaceLogo,
+  TestMuAiLogo,
 } from '~/logos';
 
 /**
@@ -65,6 +67,11 @@ export const Credits = component$(() => {
           { Logo: CodeRabbitLogo, href: 'https://coderabbit.link/valibot' },
           { Logo: AnthropicLogo, href: 'https://www.anthropic.com' },
           { Logo: OpenAILogo, href: 'https://www.openai.com' },
+          { Logo: CloudflareLogo, href: 'https://www.cloudflare.com' },
+          {
+            Logo: TestMuAiLogo,
+            href: 'https://www.testmuai.com/?utm_medium=sponsor&utm_source=valibot',
+          },
           { Logo: DigitalOceanLogo, href: 'https://www.digitalocean.com/' },
           { Logo: AlgoliaLogo, href: 'https://www.algolia.com' },
           { Logo: BuilderLogo, href: 'https://www.builder.io' },