+++ library/src/schemas/looseObject/looseObjectAsync.ts
@@ -109,19 +109,22 @@ export function looseObjectAsync(
         // parse input of key or default value asynchronously
         const valueDatasets = await Promise.all(
           Object.entries(this.entries).map(async ([key, valueSchema]) => {
+            const isKeyPresent = Object.prototype.hasOwnProperty.call(
+              input,
+              key
+            );
             if (
-              key in input ||
+              isKeyPresent ||
               ((valueSchema.type === 'exact_optional' ||
                 valueSchema.type === 'optional' ||
                 valueSchema.type === 'nullish') &&
                 // @ts-expect-error
                 valueSchema.default !== undefined)
             ) {
-              const value: unknown =
-                key in input
-                  ? // @ts-expect-error
-                    input[key]
-                  : await getDefault(valueSchema);
+              const value: unknown = isKeyPresent
+                ? // @ts-expect-error
+                  input[key]
+                : await getDefault(valueSchema);
               return [
                 key,
                 value,
@@ -129,13 +132,7 @@ export function looseObjectAsync(
                 await valueSchema['~run']({ value }, config),
               ] as const;
             }
-            return [
-              key,
-              // @ts-expect-error
-              input[key] as unknown,
-              valueSchema,
-              null,
-            ] as const;
+            return [key, undefined, valueSchema, null] as const;
           })
         );